Recover from saturation reads from ISP hist stats.
The ISP hist stats put all the data into the last bin if saturation
either low or high is encountered. We recover fine from the high/too
bright condition. For the low condition, we now force what should
be a too high condition and then hopefully the target can be achieved
with the too high code path.
If it really is too bright or too dark, there is nothing we can do.
The code will now sleep for 15 seconds after a few tries in this
condition before trying again.